+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+### [1.0.3] 15-06-2025
+
+## Added
+- User files was added.
+
+## Updated
+- Sidebar menu was updated.
+
+### [1.0.2] 15-06-2025
+
+## Added
+- Student service was added.
+
+### [1.0.1] 15-06-2025
+
+## Added
+- Model and domain service for student was added.
+
 ### [1.0.0] 15-06-2025
 
 ## Added